add clean CLI
