+++ fate_flow/apps/table_app.py
@@ -13,8 +13,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-from fate_flow.manager.tracking import Tracking
-from fate_flow.utils import job_utils
+from fate_flow.manager.data_manager import query_data_view, delete_table
 from fate_flow.utils.api_utils import get_json_result
 from fate_flow.settings import stat_logger
 from arch.api.utils.dtable_utils import get_table_info
@@ -33,7 +32,7 @@ def internal_server_error(e):
 @manager.route('/delete', methods=['post'])
 def table_delete():
     request_data = request.json
-    data_views = job_utils.query_data_view(**request_data)
+    data_views = query_data_view(**request_data)
     table_name = request_data.get('table_name')
     namespace = request_data.get('namespace')
     data = []
@@ -70,17 +69,3 @@ def dtable(table_func):
         return get_json_result()
 
 
-def delete_table(data_views):
-    data = []
-    for data_view in data_views:
-        table_name = data_view.f_table_name
-        namespace = data_view.f_table_namespace
-        table_info = {'table_name': table_name, 'namespace': namespace}
-        if table_name and namespace and table_info not in data:
-            table = session.get_data_table(name=table_name, namespace=namespace)
-            try:
-                table.destroy()
-                data.append(table_info)
-            except:
-                pass
-    return data
